more templating...
table_corplatefile.rb is a stage two template.. it is a ruby code snippet for :fields.
table_forma1cand.rb is a stage one template... it is a ruby comment thta shows what fields should exist, just fyi.
table_formb10exp.rb is a stage three template... it is a full JOBS object, all that needs to be filled in is :model.

Each of the table_ files is stage 1, 2, or, 3, depending on when it was locked/claimed by other team members.